Simple Magic round 3 slice 2: a mend can no longer wound
was: ActiveEffects.applyOnce read
  Math.max(FLOOR, Math.min(max, target.hp() + magnitude))
and the addition ran in int BEFORE either clamp could look at it. Nothing
bounds a VITALITY magnitude -- the vitality floor is exactly what is supposed
to make that safe -- so a mend authored large wrapped negative on the way in,
the clamps then clamped the wrapped number perfectly, and a crafting that
toasted "On them: +2000000000 hit points" and "THE LINK HOLDS" put the body on
the vitality floor. A heal that wounds while the screen narrates a heal is the
loudest version there is of a crafting lying about what it did.

now: the sum is taken in long, so the clamps see the real number and the sign
of the magnitude is the direction the body moves, always.

new: MendNeverWoundsTest sweeps the whole legal VITALITY range on both signs
through the table's own dose path -- 1..10, powers of two to 2^30, the short
boundaries, MAX_VALUE and MIN_VALUE -- and asserts at every step that the
landed value is exactly the clamp of the honest sum, that a positive magnitude
never takes hit points off and a negative one never puts them on, and that
both structural bounds hold. Then it casts the two extremes through the shared
verb the craftings bar drives: the int-ceiling mend arrives at FULL health
(it used to arrive on the floor) and the int-floor wound stops exactly on the
floor with the body still standing.

Verified the sweep fails against the old arithmetic before it passed against
the new: 2 of 3 red with the int addition restored.

+++ sim-core/src/main/java/com/trojia/sim/actor/spell/ActiveEffects.java
@@ -331,6 +331,16 @@ public final class ActiveEffects {
      * arriving here means a component got past the loader — a loud throw rather than the silent
      * return this used to do, which is exactly how five pairings came to resolve, charge a
      * resist, narrate success and change nothing.
+     *
+     * <p><b>The sum is taken in {@code long}, and that is load-bearing.</b> This used to read
+     * {@code Math.max(FLOOR, Math.min(max, target.hp() + magnitude))} — and the addition ran in
+     * {@code int} BEFORE either clamp could see it. Nothing bounds a VITALITY magnitude (the
+     * floor below is what makes that safe), so a mend authored large wrapped negative on the way
+     * in, the clamps then faithfully clamped the wrapped number, and a crafting that narrated
+     * "the link holds" and printed "On them: +2000000000 hit points" delivered the maximum wound
+     * the floor allows. A heal that wounds is the loudest possible version of a crafting lying
+     * about what it did. In {@code long} the clamps see the real number, so the sign of the
+     * magnitude is the direction the body moves, always.
      */
     static void applyOnce(Actor target, EffectKind kind, int param, int magnitude) {
         if (kind != EffectKind.VITALITY) {
@@ -338,8 +348,8 @@ public final class ActiveEffects {
                     + " -- EffectPairing refuses that pairing at load, so reaching applyOnce with"
                     + " it means a component was built without its canonical constructor");
         }
-        int max = target.stats().hp();
-        int next = Math.max(VITALITY_FLOOR, Math.min(max, target.hp() + magnitude));
+        long max = target.stats().hp();
+        long next = Math.max(VITALITY_FLOOR, Math.min(max, (long) target.hp() + magnitude));
         target.setHp((short) next);
     }
